v2

+++ app/Http/Controllers/PackageController.php
@@ -12,6 +12,7 @@ use App\models\PackageActivityRelations;
 use App\models\PackagePic;
 use App\models\PackageTag;
 use App\models\PackageTagRelation;
+use App\models\PackageThumbnailsPic;
 use App\models\Tags;
 use Illuminate\Http\Request;
 
@@ -69,6 +70,11 @@ class PackageController extends Controller
         foreach ($destinations as $key => $item)
             $item->category = DestinationCategory::find($key);
 
+        $package->thumbnail = PackageThumbnailsPic::where('packageId', $package->id)->get();
+        foreach ($package->thumbnail as $item)
+            $item->pic = asset('uploaded/packages/' . $package->id . '/thumbnail_' . $item->pic);
+
+
         $activity = Activity::all();
 
         $allDestination = Destination::all();
@@ -77,12 +83,18 @@ class PackageController extends Controller
 
     public function storePackage(Request $request)
     {
-        if(isset($request->name) && isset($request->id) && isset($request->lat) && isset($request->lng) && isset($request->destinationId) && isset($request->mainActivity)){
+        if(isset($request->name) && isset($request->id) && isset($request->lat) && isset($request->lng) && isset($request->destinationId) && isset($request->code) && isset($request->mainActivity)){
 
             if($request->id == 0){
                 $pack = Package::where('name', $request->name)->where('destId', $request->destinationId)->first();
                 if($pack != null) {
                     echo json_encode(['status' => 'nok2']);
+                    return;
+                }
+
+                $codePack = Package::where('code', $request->code)->first();
+                if($codePack != null) {
+                    echo json_encode(['status' => 'nok9']);
                     return;
                 }
 
@@ -92,6 +104,12 @@ class PackageController extends Controller
                 $pack = Package::where('name', $request->name)->where('destId', $request->destinationId)->where('id' , '!=', $request->id)->first();
                 if($pack != null) {
                     echo json_encode(['status' => 'nok2']);
+                    return;
+                }
+
+                $codePack = Package::where('code', $request->code)->where('id' , '!=', $request->id)->first();
+                if($codePack != null) {
+                    echo json_encode(['status' => 'nok9']);
                     return;
                 }
 
@@ -109,10 +127,12 @@ class PackageController extends Controller
             $pack->lat = $request->lat;
             $pack->lng = $request->lng;
             $pack->day = $request->day;
+            $pack->code = $request->code;
             $pack->season = $request->season;
             $pack->sDate = $request->sDate;
             $pack->eDate = $request->eDate;
             $pack->money = $request->cost;
+            $pack->level = $request->level;
             $pack->showPack = $request->showPack;
             $pack->mainActivityId = $request->mainActivity;
             $pack->save();
@@ -167,9 +187,48 @@ class PackageController extends Controller
                 $location .= '/' . $request->id;
                 if(!file_exists($location))
                     mkdir($location);
+
+                $image = $request->file('pic');
+                $dirs = 'uploaded/packages/' . $request->id;
+
+                if($request->kind == 'thumbnail'){
+                    $size = [
+                        [
+                            'width' => null,
+                            'height' => 100,
+                            'name' => 'thumbnail_',
+                            'destination' => $dirs
+                        ]
+                    ];
+                }
+                else{
+                    $size = [
+                        [
+                            'width' => null,
+                            'height' => 250,
+                            'name' => 'min_',
+                            'destination' => $dirs
+                        ],
+                        [
+                            'width' => null,
+                            'height' => 400,
+                            'name' => 'list_',
+                            'destination' => $dirs
+                        ],
+                        [
+                            'width' => null,
+                            'height' => 500,
+                            'name' => 'slide_',
+                            'destination' => $dirs
+                        ]
+                    ];
+                }
+
+                $fileName = resizeImage($image, $size);
+
                 $location .= '/' . $fileName;
 
-                $picResult = compressImage($_FILES['pic']['tmp_name'], $location, 80);
+                $picResult = storeImage($_FILES['pic']['tmp_name'], $location);
                 if ($picResult) {
                     if($request->kind == 'mainPic') {
                         if ($pack->pic != null)
@@ -177,6 +236,14 @@ class PackageController extends Controller
                         $pack->pic = $fileName;
                         $pack->save();
                         echo json_encode(['status' => 'ok']);
+                    }
+                    else if($request->kind == 'thumbnail'){
+                        $thumbnail = new PackageThumbnailsPic();
+                        $thumbnail->packageId = $pack->id;
+                        $thumbnail->pic = $fileName;
+                        $thumbnail->save();
+
+                        echo json_encode(['status' => 'ok', 'id' => $thumbnail->id]);
                     }
                     else{
                         $sidePic = new PackagePic();
@@ -202,15 +269,30 @@ class PackageController extends Controller
 
     public function deleteImgPackage(Request $request)
     {
-        if(isset($request->id)){
-            $pic = PackagePic::find($request->id);
-            if($pic != null){
-                \File::delete('uploaded/packages/' . $pic->packageId . '/' . $pic->pic);
-                $pic->delete();
-                echo json_encode(['status' => 'ok']);
+        if(isset($request->id) && isset($request->kind)){
+
+            if($request->kind == 'side') {
+                $pic = PackagePic::find($request->id);
+                if ($pic != null) {
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/slide_' . $pic->pic);
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/list_' . $pic->pic);
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/min_' . $pic->pic);
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/' . $pic->pic);
+                    $pic->delete();
+                    echo json_encode(['status' => 'ok']);
+                } else
+                    echo json_encode(['status' => 'nok1']);
             }
-            else
-                echo json_encode(['status' => 'nok1']);
+            else{
+                $pic = PackageThumbnailsPic::find($request->id);
+                if ($pic != null) {
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/thumbnail_' . $pic->pic);
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/' . $pic->pic);
+                    $pic->delete();
+                    echo json_encode(['status' => 'ok']);
+                } else
+                    echo json_encode(['status' => 'nok1']);
+            }
         }
         else
             echo json_encode(['status' => 'nok']);
@@ -227,13 +309,82 @@ class PackageController extends Controller
                 PackageActivityRelations::where('packageId', $package->id)->delete();
                 $pics = PackagePic::where('packageId', $package->id)->get();
                 foreach ($pics as $pic){
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/slide_' . $pic->pic);
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/min_' . $pic->pic);
+                    \File::delete('uploaded/packages/' . $pic->packageId . '/list_' . $pic->pic);
                     \File::delete('uploaded/packages/' . $pic->packageId . '/' . $pic->pic);
                     $pic->delete();
                 }
 
+                \File::delete('uploaded/packages/' . $package->id . '/slide_' . $package->pic);
+                \File::delete('uploaded/packages/' . $package->id . '/min_' . $package->pic);
+                \File::delete('uploaded/packages/' . $package->id . '/list_' . $package->pic);
                 \File::delete('uploaded/packages/' . $package->id . '/' . $package->pic);
                 $package->delete();
                 echo json_encode(['status' => 'ok']);
+            }
+            else
+                echo json_encode(['status' => 'nok1']);
+        }
+        else
+            echo json_encode(['status' => 'nok']);
+
+        return;
+    }
+
+    public function storeVideoAudioPackage(Request $request)
+    {
+        if(isset($request->id) && $_FILES['file'] && $_FILES['file']['error'] == 0){
+            $packe = Package::find($request->id);
+            if($packe != null){
+                $fileName = $_FILES['file']['name'];
+                $location = __DIR__ . '/../../../public/uploaded/packages';
+                if(!file_exists($location))
+                    mkdir($location);
+                $location .= '/' . $request->id;
+                if(!file_exists($location))
+                    mkdir($location);
+
+                if(is_file($location . '/' . $fileName))
+                    $fileName = time() . $fileName;
+
+                $location .= '/' . $fileName;
+
+                if(move_uploaded_file($_FILES['file']['tmp_name'], $location)) {
+                    if($request->kind == 'audio'){
+                        if($packe->podcast != null)
+                            \File::delete('uploaded/packages/'. $packe->id . '/' . $packe->podcast);
+
+                        $packe->podcast = $fileName;
+                        $packe->save();
+
+                        $videoUrl = asset('uploaded/packages/' . $packe->id . '/' . $packe->podcast);
+                        echo json_encode(['status' => 'ok', 'result' => $videoUrl]);
+                    }
+                    else if($request->kind == 'video'){
+                        if($packe->video != null)
+                            \File::delete('uploaded/packages/'. $packe->id . '/' . $packe->video);
+
+                        $packe->video = $fileName;
+                        $packe->save();
+
+                        $videoUrl = asset('uploaded/packages/' . $packe->id . '/' . $packe->video);
+                        echo json_encode(['status' => 'ok', 'result' => $videoUrl]);
+                    }
+                    else if($request->kind == 'brochure'){
+                        if($packe->brochure != null)
+                            \File::delete('uploaded/packages/'. $packe->id . '/' . $packe->brochure);
+
+                        $packe->brochure = $fileName;
+                        $packe->save();
+
+                        $brochureUrl = asset('uploaded/packages/' . $packe->id . '/' . $packe->brochure);
+                        echo json_encode(['status' => 'ok', 'url' => $brochureUrl, 'name' => $fileName]);
+                    }
+
+                }
+                else
+                    echo json_encode(['status' => 'nok2']);
             }
             else
                 echo json_encode(['status' => 'nok1']);
